Move lexer and parser files to language submodule and remove test file.
