+++ maltoolbox/language/lexer_parser/.gitignore
@@ -1,13 +1,11 @@
-test_langs/
 .venv/
 
 # antlr generated files we don't need
 malVisitor.py
 malListener.py
+malLexer.py
+malParser.py
 mal.interp
 mal.tokens
 malLexer.interp
 malLexer.tokens
-
-
-*.zip
