+++ maltoolbox/language/lexer_parser/mal_lexer.py
